diffgui: handle some corner case
When the two bom have different top node,
a 3rd top node has to be created, and set as the root of
the two bom.

corner case #1: Depending by the origin of the
bom this top node may be an integer or a str. Check the type
and create a new top node accordling.

corner case #2: when the a bom receive a new top node, it is possible
that there are some references (e.g. "deps")  to the old top node (e.g.
when a bom is imported). Change the refrence too.