Allow the bombrowser.ini to be in another place
Allow the bombrowser.init file to be in another place. This plase
must be stored in the file 'bombrowser.forward'.

Removed the (unused) capability to store the configuration in the
DB, due to the complexity and to avoid the egg/chicken problem:
a) load the config
b) open the db, with the setting in the config
c) reload again the config from the DB

Clean up the several step of opening the config the check the download
from db the recheck again....

Removed the unused function db.get_config() and the related test.